Add navigator markup tests
