+++ testing/tests/DevExpress.ui.widgets.scheduler/navigator.markup.tests.js
@@ -1,0 +1,252 @@
+"use strict";
+
+import $ from "jquery";
+import SchedulerNavigator from "ui/scheduler/ui.scheduler.navigator";
+import dateLocalization from "localization/date";
+import devices from "core/devices";
+import noop from "core/utils/common";
+import "ui/scheduler/ui.scheduler";
+
+"use strict";
+
+const moduleConfig = {
+    beforeEach: () => {
+        devices.current({ platform: "generic" });
+        this.instance = $("#navigator").dxSchedulerNavigator().dxSchedulerNavigator("instance");
+        this.instance.notifyObserver = noop.noop;
+    }
+};
+
+QUnit.module("Navigator markup", moduleConfig, () => {
+    QUnit.test("Scheduler navigator should be initialized", (assert) => {
+        assert.ok(this.instance instanceof SchedulerNavigator, "dxSchedulerNavigator was initialized");
+    });
+
+    QUnit.test("Scheduler navigator should have a right css class", (assert) => {
+        var $element = this.instance.$element();
+
+        assert.ok($element.hasClass("dx-scheduler-navigator"), "dxSchedulerNavigator has 'dx-scheduler-navigator' css class");
+    });
+
+    QUnit.test("Scheduler navigator should contain the 'next' button", (assert) => {
+        var $element = this.instance.$element();
+
+        var $button = $element.find(".dx-scheduler-navigator-next");
+
+        assert.equal($button.length, 1, "Navigator 'next' is rendered");
+        assert.ok($button.hasClass("dx-button"), "Navigator 'next' is dxButton");
+        assert.equal($button.attr("aria-label"), "Next period", "Navigator 'next' button has right label");
+    });
+
+    QUnit.test("Scheduler navigator should contain the 'previous' button", (assert) => {
+        var $element = this.instance.$element();
+
+        var $button = $element.find(".dx-scheduler-navigator-previous");
+
+        assert.equal($button.length, 1, "Navigator 'previous' is rendered");
+        assert.ok($button.hasClass("dx-button"), "Navigator 'previous' is dxButton");
+        assert.equal($button.attr("aria-label"), "Previous period", "Navigator 'previous' button has right label");
+    });
+
+    QUnit.test("Scheduler navigator should contain the 'caption' button", (assert) => {
+        var $element = this.instance.$element();
+
+        var $button = $element.find(".dx-scheduler-navigator-caption");
+
+        assert.equal($button.length, 1, "Navigation 'caption' is rendered");
+        assert.ok($button.hasClass("dx-button"), "Navigator 'caption' is dxButton");
+    });
+
+    QUnit.test("Caption should be OK with default options", (assert) => {
+        var $element = this.instance.$element(),
+            button = $element.find(".dx-scheduler-navigator-caption").dxButton("instance"),
+            date = new Date(),
+            caption = [dateLocalization.format(date, "day"), dateLocalization.format(date, "monthAndYear")].join(" ");
+
+        assert.equal(button.option("text"), caption, "Caption is OK");
+    });
+
+    QUnit.test("Caption should be OK when step and date are changed", (assert) => {
+        var $element = this.instance.$element(),
+            button = $element.find(".dx-scheduler-navigator-caption").dxButton("instance"),
+            date = new Date(2015, 0, 24),
+            caption = "24 January 2015";
+
+        this.instance.option("firstDayOfWeek", 1);
+        this.instance.option("date", date);
+        assert.equal(button.option("text"), caption, "Step is day: Caption is OK");
+
+        var weekDateString = devices.real().generic ? "25 January 2015" : "25 Jan 2015";
+
+        caption = "19-" + weekDateString;
+
+        this.instance.option("step", "week");
+        assert.equal(button.option("text"), caption, "Step is week: Caption is OK");
+
+        weekDateString = devices.real().generic ? "23 January 2015" : "23 Jan 2015",
+        caption = "19-" + weekDateString;
+
+        this.instance.option("step", "workWeek");
+        assert.equal(button.option("text"), caption, "Step is workWeek: Caption is OK");
+
+        caption = dateLocalization.format(date, "monthandyear");
+        this.instance.option("step", "month");
+        assert.equal(button.option("text"), caption, "Step is month: Caption is OK");
+
+
+        this.instance.option("step", "agenda");
+
+        weekDateString = devices.real().generic ? "30 January 2015" : "30 Jan 2015";
+        caption = "24-" + weekDateString;
+
+        assert.equal(button.option("text"), caption, "Step is agenda: Caption is OK");
+    });
+
+    QUnit.test("Caption should be OK for workWeek view & firstDayOfWeek = 0", (assert) => {
+        var $element = this.instance.$element(),
+            button = $element.find(".dx-scheduler-navigator-caption").dxButton("instance"),
+            date = new Date(2015, 0, 24);
+
+        this.instance.option("firstDayOfWeek", 0);
+        this.instance.option("date", date);
+
+        var caption = devices.real().generic ? "19-23 January 2015" : "19-23 Jan 2015";
+
+        this.instance.option("step", "workWeek");
+        assert.equal(button.option("text"), caption, "Step is workWeek: Caption is OK");
+    });
+
+    QUnit.test("Caption should be OK for Day with intervalCount", (assert) => {
+        var $element = this.instance.$element(),
+            button = $element.find(".dx-scheduler-navigator-caption").dxButton("instance"),
+            date = new Date(2015, 4, 25),
+            caption = "25-27 May 2015";
+
+        this.instance.option("date", date);
+        this.instance.option("intervalCount", 3);
+
+        assert.equal(button.option("text"), caption, "Caption is OK");
+    });
+
+    QUnit.test("Caption should be OK for workWeek view with intervalCount", (assert) => {
+        var $element = this.instance.$element(),
+            button = $element.find(".dx-scheduler-navigator-caption").dxButton("instance"),
+            date = new Date(2015, 4, 25),
+            caption = "25 May-12 Jun 2015";
+
+        this.instance.option("firstDayOfWeek", 1);
+        this.instance.option("date", date);
+        this.instance.option("intervalCount", 3),
+        this.instance.option("step", "workWeek");
+
+        assert.equal(button.option("text"), caption, "Caption is OK");
+    });
+
+    QUnit.test("Caption should be OK for week view with intervalCount", (assert) => {
+        var $element = this.instance.$element(),
+            button = $element.find(".dx-scheduler-navigator-caption").dxButton("instance"),
+            date = new Date(2015, 4, 25),
+            caption = "25 May-14 Jun 2015";
+
+        this.instance.option("firstDayOfWeek", 1);
+        this.instance.option("date", date);
+        this.instance.option("intervalCount", 3),
+        this.instance.option("step", "week");
+
+        assert.equal(button.option("text"), caption, "Caption is OK");
+    });
+
+    QUnit.test("Caption should be OK for Month with intervalCount", (assert) => {
+        var $element = this.instance.$element(),
+            button = $element.find(".dx-scheduler-navigator-caption").dxButton("instance"),
+            date = new Date(2017, 4, 25),
+            caption = "Jun-Jul 2017";
+
+        this.instance.option("date", date);
+        this.instance.option("intervalCount", 2);
+        this.instance.option("step", "month");
+
+        assert.equal(button.option("text"), caption, "Caption is OK");
+    });
+
+    QUnit.test("Caption should be OK for Month with intervalCount for different years", (assert) => {
+        var $element = this.instance.$element(),
+            button = $element.find(".dx-scheduler-navigator-caption").dxButton("instance"),
+            date = new Date(2017, 10, 25),
+            caption = "Dec 2017-Feb 2018";
+
+        this.instance.option("date", date);
+        this.instance.option("intervalCount", 3);
+        this.instance.option("step", "month");
+
+        assert.equal(button.option("text"), caption, "Caption is OK");
+    });
+
+    QUnit.test("Caption should be OK for workWeek view, if date = Sunday", (assert) => {
+        var $element = this.instance.$element(),
+            button = $element.find(".dx-scheduler-navigator-caption").dxButton("instance"),
+            date = new Date(2016, 0, 10);
+
+        this.instance.option("firstDayOfWeek", 0);
+        this.instance.option("date", date);
+
+        var caption = devices.real().generic ? "11-15 January 2016" : "11-15 Jan 2016";
+
+        this.instance.option("step", "workWeek");
+        assert.equal(button.option("text"), caption, "Step is workWeek: Caption is OK");
+    });
+
+    QUnit.test("Caption should be OK for workWeek view, if date = Sunday and firstDayOfWeek != 0", (assert) => {
+        var $element = this.instance.$element(),
+            button = $element.find(".dx-scheduler-navigator-caption").dxButton("instance"),
+            date = new Date(2016, 0, 10);
+
+        this.instance.option("firstDayOfWeek", 3);
+        this.instance.option("date", date);
+
+        var caption = devices.real().generic ? "6-12 January 2016" : "6-12 Jan 2016";
+
+        this.instance.option("step", "workWeek");
+        assert.equal(button.option("text"), caption, "Step is workWeek: Caption is OK");
+    });
+
+    QUnit.test("Caption should be OK for workWeek view, if date = Saturday & firstDayOfWeek = 6", (assert) => {
+        var $element = this.instance.$element(),
+            button = $element.find(".dx-scheduler-navigator-caption").dxButton("instance"),
+            date = new Date(2016, 0, 9);
+
+        this.instance.option("firstDayOfWeek", 6);
+        this.instance.option("date", date);
+
+        var caption = devices.real().generic ? "11-15 January 2016" : "11-15 Jan 2016";
+
+        this.instance.option("step", "workWeek");
+        assert.equal(button.option("text"), caption, "Step is workWeek: Caption is OK");
+    });
+
+    QUnit.test("Caption should be OK for workWeek view, if date = Sunday & firstDayOfWeek = 6", (assert) => {
+        var $element = this.instance.$element(),
+            button = $element.find(".dx-scheduler-navigator-caption").dxButton("instance"),
+            date = new Date(2016, 0, 10);
+
+        this.instance.option("firstDayOfWeek", 6);
+        this.instance.option("date", date);
+
+        var caption = devices.real().generic ? "11-15 January 2016" : "11-15 Jan 2016";
+
+        this.instance.option("step", "workWeek");
+        assert.equal(button.option("text"), caption, "Step is workWeek: Caption is OK");
+    });
+
+    QUnit.test("Caption should be OK for week view, if date = Sunday & firstDayOfWeek = 1", (assert) => {
+        var $element = this.instance.$element(),
+            button = $element.find(".dx-scheduler-navigator-caption").dxButton("instance"),
+            date = new Date(2015, 2, 1);
+
+        this.instance.option("firstDayOfWeek", 1);
+        this.instance.option("date", date);
+
+        this.instance.option("step", "week");
+        assert.equal(button.option("text"), "23 Feb-1 Mar 2015", "Step is week: Caption is OK");
+    });
+});
