+++ testing/tests/DevExpress.ui.widgets.scheduler/scheduler.markup.tests.js
@@ -1,4 +1,16 @@
 "use strict";
 
-require("./workSpace.markup.tests.js");
-require("./common.markup.tests.js");
+import $ from "jquery";
+
+QUnit.testStart(() => {
+    const markup =
+        '<div id="scheduler"> </div>\
+        <div id="scheduler-work-space"> </div>\
+        <div id="navigator"> </div>';
+
+    $("#qunit-fixture").html(markup);
+});
+
+import "./workSpace.markup.tests.js";
+import "./common.markup.tests.js";
+import "./navigator.markup.tests.js";
